+++ pages/[streamer].js
@@ -7,22 +7,9 @@ import StreamerAvatar from "@components/StreamerAvatar";
 import StreamerSchedule from "@components/StreamerSchedule";
 import LatestStream from "@components/LatestStream";
 import Twitch from "@components/Svg/Twitch";
-import { transformEmotes } from "@utils/Tools";
+import { transformEmotes, getEmotesForOg } from "@utils/Tools";
 import { IMG_WIDTH, IMG_HEIGHT, generateImageUrl } from "@utils/OpenGraph";
-import { getRandomEntry } from "@whitep4nth3r/get-random-entry";
 
-function getEmotesForOg(emotes) {
-  const emoteGroup = new Set();
-
-  const returnArraySize = emotes.length >= 5 ? 5 : emotes.length;
-
-  while (emoteGroup.size < returnArraySize) {
-    let newRandomEntry = getRandomEntry(emotes);
-    emoteGroup.add(newRandomEntry.imageUrl);
-  }
-
-  return Array.from(emoteGroup);
-}
 
 export default function Streamer({ streamer, hasBanner }) {
   const transformedEmotes = transformEmotes(streamer.emotes);
