+++ utils/Tools.js
@@ -1,3 +1,5 @@
+import { getRandomEntry } from "@whitep4nth3r/get-random-entry";
+
 export function transformEmotes(emotes) {
   return emotes?.length > 0
     ? emotes.map((emote) => {
@@ -7,4 +9,17 @@ export function transformEmotes(emotes) {
         };
       })
     : [];
+}
+
+export function getEmotesForOg(emotes) {
+  const emoteGroup = new Set();
+
+  const returnArraySize = emotes.length >= 5 ? 5 : emotes.length;
+
+  while (emoteGroup.size < returnArraySize) {
+    let newRandomEntry = getRandomEntry(emotes);
+    emoteGroup.add(newRandomEntry.imageUrl);
+  }
+
+  return Array.from(emoteGroup);
 }
